% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_distinct.R
\name{col_distinct}
\alias{col_distinct}
\title{Check colours are perceptually distinct}
\usage{
col_distinct(
  colours,
  threshold = 20,
  method = "cmc",
  lightness = 1,
  cvd = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{colours}{a colour vector.}

\item{threshold}{cutoff for clustering colours; meaning depends on \code{method}.}

\item{method}{distance method: see \code{\link[farver:compare_colour]{farver::compare_colour()}}.}

\item{lightness, ...}{other params passed to \code{\link[farver:compare_colour]{farver::compare_colour()}}.}

\item{cvd}{if \code{TRUE} simular colour vision disorders.
Requires \code{colorBlindness} to be installed.}

\item{plot}{if \code{TRUE} plot a dendrogram of the colour clusters.
Names from \code{colours} are used to label the horixontal axis.
Boxes are drawn around each cluster and coloured by the representative.}
}
\value{
vector of distinct colours from \code{colours}
}
\description{
Some palettes can include colours that are visually similar, especially to
people with colour vision disorders. This utility removes overly similar
colours in a set. This is done by clustering using the specified \code{method}
and \code{threshold}. It can be useful to set \code{plot = TRUE} to decide on a
suitable value of \code{threshold}.
}
\examples{

col_nsw(1:10, v = "b") |> col_distinct(plot = TRUE)
col_nsw(1:10, v = "b") |> col_distinct(plot = TRUE, cvd = TRUE)
}
