% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_nsw.R
\docType{data}
\name{nsw_colours}
\alias{nsw_colours}
\alias{col_nsw}
\title{Access NSW palette grids by rows and/or columns}
\usage{
nsw_colours

col_nsw(
  hue,
  tone,
  variant = getOption("waratah.colour_theme", default = "base"),
  byrow = FALSE
)
}
\arguments{
\item{hue}{Name or index of the hue - see below.}

\item{tone}{Name or index of the tone - see below.}

\item{variant}{Name of palette variant.
Available options are: base, aboriginal, corporate, treasury.}

\item{byrow}{Vary \code{tone} faster than \code{hue} if \code{TRUE}.}
}
\value{
Named vector of colours
}
\description{
The NSW design system colours work in grids of colour columns and tonal rows.
This function allows accessing a colour grid like a matrix.
For ggplot colour palettes, you'll normally want \code{\link[=pal_nsw]{pal_nsw()}} instead.
}
\details{
\code{hue} and \code{tone} work the same way as matrix indexing with \code{[} in that they
can be used to return single or multiple entries from the grid.
They can be character vectors, integer vectors, or logicals.

Anchor colours used to create the NSW colour palettes can also be used
stand-alone (e.g. \code{nsw_colours$blue_01} is \code{"#002664"}).
}
\section{Colour columns and tonal rows}{

The \code{"base"} variant supports:
\itemize{
\item \strong{hue}: greys, greens, teals, blues, purples, fucshias, reds, oranges, yellows, browns
\item \strong{tone}: dark, normal, light, pale
}

The \code{"aboriginal"} variant supports:
\itemize{
\item \strong{hue}: reds, oranges, browns, yellows, greens, blues, purples, greys
\item \strong{tone}: dark, normal, light, pale
}

Other variants (treasury, corporate)
are colour themes, which are subsets of the main two grids using only certain
colour columns.

Unambiguous shortened forms are accepted, e.g. \code{pal_nsw(h = "red", v = "a")}.
}

\examples{
col_nsw(h = "blue", v = "aboriginal")
col_nsw(h = c("teal", "orange"), t = 1:2)
col_nsw(h = c("teal", "orange"), t = 1:2, byrow = TRUE)

}
\seealso{
\code{\link[=pal_nsw]{pal_nsw()}}
}
\keyword{datasets}
