% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsw_colours.R
\docType{data}
\name{nsw_colours}
\alias{nsw_colours}
\title{NSW Design System Colours}
\format{
An object of class \code{list} of length 75.
}
\usage{
nsw_colours
}
\description{
Anchor colours used to create the waratah colour palettes. These can also be used as
stand-alone colours (e.g. waratah_colours$blue_01 returns the hex code \"#002664\")
}
\details{
Colours are available to view at
https://digitalnsw.github.io/nsw-design-system/docs/content/design/theming.html
}
\keyword{datasets}
