% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_waratah.R
\name{scale_colour_waratah}
\alias{scale_colour_waratah}
\alias{scale_fill_waratah}
\alias{scale_colour_waratah_c}
\alias{scale_fill_waratah_c}
\title{Waratah colour scales for 'ggplot2'}
\usage{
scale_colour_waratah(
  name = ggplot2::waiver(),
  ...,
  palette = "default",
  direction = 1,
  aesthetics = "colour"
)

scale_fill_waratah(
  name = ggplot2::waiver(),
  ...,
  palette = "default",
  direction = 1,
  aesthetics = "fill"
)

scale_colour_waratah_c(
  name = ggplot2::waiver(),
  ...,
  palette = "default",
  direction = -1,
  values = NULL,
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "colour"
)

scale_fill_waratah_c(
  name = ggplot2::waiver(),
  ...,
  palette = "default",
  direction = -1,
  values = NULL,
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill"
)
}
\arguments{
\item{name, na.value, guide, aesthetics, ...}{Arguments passed on to
\code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}} to control name,
limits, breaks, labels and so forth.}

\item{palette}{Name of the palette: default, brand_default, neg_to_pos, greens, teals, blues, purples, fuchsias, reds, oranges, yellows, greys, browns, deeps, lights, warm_colours, cool_colours, colour_blind}

\item{direction}{Set to -1 to reverse the order of colours in the palette,
or 1 for the original order.}

\item{values}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each anchor colour.}
}
\description{
These scales can be used with ggplot to use the waratah palettes. The \verb{_c}
variants interpolate the palette colours into a continuous gradient.
}
