% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_waratah.R
\name{theme_waratah}
\alias{theme_waratah}
\title{Theme plots using NSW colours and fonts}
\usage{
theme_waratah(
  base_size = 11,
  base_family = "Arial",
  header_family = "Public Sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ink = "black",
  paper = "off_white",
  accent = "blue_01",
  show_grid_lines = TRUE,
  void = FALSE
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{header_family}{font family for titles and headers. The default, \code{NULL},
uses theme inheritance to set the font. This setting affects axis titles,
legend titles, the plot title and tag text.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{ink, paper, accent}{colour for foreground, background, and accented elements respectively.}

\item{show_grid_lines}{whether to show grid lines.
If \code{FALSE}, all grid lines are removed but the axis text is retained.
Ignored when \code{void} is \code{TRUE}.}

\item{void}{whether to hide grid lines and axes.
If \code{TRUE}, all grid lines and axes are removed. This is useful when creating
pie/donut charts.}
}
\value{
ggplot theme specification to add to a plot
}
\description{
A 'ggplot2' theme compatible with the NSW design system.
It sets default colour scales, fonts, and some other styles.
}
\examples{
library(ggplot2)
set_theme(theme_waratah())

ggplot(palmerpenguins::penguins) +
  geom_point(aes(
    x = bill_length_mm,
    y = flipper_length_mm,
    colour = species,
    size = body_mass_g
  )) +
  labs(
    title = "Let's try some *italics* in the title",
    subtitle = paste0(
      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do",
      " eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut",
      " enim ad minim veniam, quis nostrud exercitation ullamco laboris",
      " nisi ut aliquip ex ea commodo consequat."
    ),
    caption = "Data from {palmerpenguins}",
    dictionary = c(
      bill_length_mm = "Bill length (mm)",
      flipper_length_mm = "Flipper length (mm)",
      species = "Species"
    )
  ) +
  scale_size(
    "Body mass (kg)",
    labels = scales::label_number(scale = 1/1000)
  )

}
\examples{
library(ggplot2)

p1 <-
  ggplot(palmerpenguins::penguins) +
  geom_point(aes(
    x = bill_length_mm,
    y = flipper_length_mm,
    colour = species,
    size = body_mass_g
  )) +
  labs(
    dictionary = c(
      bill_length_mm = "Bill length (mm)",
      flipper_length_mm = "Flipper length (mm)",
      species = "Species",
      body_mass_g = "Body mass (g)"
    ),
    caption = "Data from {palmerpenguins}"
  ) +
  theme_waratah()

}
