% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waratah_colours.R
\docType{data}
\name{waratah_colours}
\alias{waratah_colours}
\title{waratah_colours}
\format{
A list with 47 elements
}
\usage{
waratah_colours
}
\description{
Anchor colours used to create the waratah colour palettes. These can also be used as stand-alone colours (e.g. waratah_colours$blue_01 returns the hex code \"#002664\")
}
\details{
Colours are available to view at https://digitalnsw.github.io/nsw-design-system/docs/content/design/theming.html
}
\keyword{datasets}
