% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{pal_interleave}
\alias{pal_interleave}
\alias{pal_stretch}
\alias{col_contrasting}
\title{Construct palette variants}
\usage{
pal_interleave(...)

pal_stretch(pal)

col_contrasting(bg)
}
\arguments{
\item{...}{two or more vectors of colours}

\item{pal}{palette object}

\item{bg}{vector of colours}
}
\value{
\itemize{
\item for \code{col_contrasting()} a vector of colours the same length as \code{bg},
\item for \verb{pal_*()} a palette object.
}
}
\description{
Depending on the structure of your data you may wish to combine palettes
according to some pattern. These helpers may come in handy.
\code{pal_interleave()} is for grouped data.
\code{pal_stretch()} interpolates a palette into a new discrete palette.
\code{col_contrasting()} chooses colours based on the given background colours.
}
