% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_waratah.R
\name{pal_waratah}
\alias{pal_waratah}
\title{Flexible colour palettes}
\usage{
pal_waratah(
  type = c("qual", "seq", "div", "pairs", "triples"),
  hue = 1,
  cvd = FALSE,
  variant = getOption("waratah.colour_theme", default = "base"),
  direction = 1
)
}
\arguments{
\item{type}{type of palette that should be generated:
\itemize{
\item \code{"qual"} discrete qualitative palette
\item \code{"seq"} continuous sequential palette: a gradient spanning extreme tones of \code{hue}
\item \code{"div"} continuous diverging palette: a gradient spanning extreme tones of \code{hue}
and a second distinct colour.
\item \code{"pairs"}, \code{"triples"} like `"qual"`` but in sets of 2 or 2 tones of each hue.
}}

\item{hue}{main hue, either by name of index.
See \code{\link[=col_nsw]{col_nsw()}} for supported values.
Only used when \code{type} is \code{"seq"} or \code{"div"}.}

\item{cvd}{when \code{TRUE} account for colour vision disorders.
Requires the \code{colorBlindness} package.}

\item{variant}{name of palette variant.
Available options are: base, aboriginal, corporate, treasury.
Ignored unless \code{hue} or \code{tone} is specified.}

\item{direction}{set to -1 to reverse the order of colours in the palette,
or 1 for the original order.}
}
\value{
A palette object (see \link[scales:new_continuous_palette]{palette constructors})
}
\description{
Unlike \code{\link[=pal_nsw]{pal_nsw()}}, \code{pal_waratah()} is not based strictly on the NSW colour
palette grids.
It tries to choose colours that are perceptually distinct, optionally taking
into account colour-blindness when doing so.
}
\examples{
library(scales)

pal_waratah("qual") |> show_col()
pal_waratah("pairs") |> show_col()
pal_waratah("seq", hue = "red") |> show_col(labels = FALSE)
pal_waratah("div", hue = "yellow", variant = "aboriginal") |> show_col(labels = FALSE)

}
\seealso{
Other palettes: 
\code{\link{pal_nsw}()}
}
\concept{palettes}
