% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_theme.R
\name{reactable_theme}
\alias{reactable_theme}
\title{reactable_theme()}
\usage{
reactable_theme(
  colour = "blue_01",
  title_font = "Public Sans",
  base_font = "Public Sans",
  base_text_size = 10,
  textColor = NULL,
  .colours = nsw_colours,
  backgroundColor = TRUE,
  borderColor = TRUE,
  stripedColor = NULL,
  ...
)
}
\arguments{
\item{colour}{The starting colour for the table.
It can be any of the ophelia_colours (e.g. "blue_01", "green_01", "purple_02").}

\item{title_font}{The default title is "Public Sans".
To change to a different font (e.g. to match up with an external publication
requirement), simply change as desired (e.g. "Times New Roman"). Make sure you've
installed the fonts you want to use on your device first!}

\item{base_font}{The default font is also "Public Sans".
If you want to use a different font (e.g. to match up with an external publication
requirement), simply change as desired (e.g. "Arial"). Make sure you've installed
the fonts you want to use on your device first!}

\item{base_text_size}{Base text size.
The header size is relative to the base text size.}

\item{textColor}{You can specify a text colour if you would like it
different to the borderColor and stripedColor.}

\item{.colours}{Used to read in nsw_colours.
Please leave as is!}

\item{backgroundColor}{Logical. Can be specified as TRUE/FALSE.
The colour is off_white #F2F2F2}

\item{borderColor}{The colour for the horizontal lines between rows.}

\item{stripedColor}{The fill colour for alternate rows.}

\item{...}{Other parameters to pass to reactable::reactableTheme().
See ?reactable::reactableTheme for documentation.
For examples of what can be done with reactable, see
https://glin.github.io/reactable/articles/cookbook/cookbook.html.}
}
\value{
A theme to apply to a reactable object
}
\description{
reactable_theme()
}
\examples{

library(reactable)
head(palmerpenguins::penguins, 10) |>
  reactable(theme = reactableTheme())

# Adding NSW style:
head(palmerpenguins::penguins, 10) |>
  reactable(theme = reactable_theme())

# Further arguments for styling
head(palmerpenguins::penguins, 10) |>
  reactable(
    theme = reactable_theme(
      colour = "blue_01",
      base_font = "Arial",
      textColor = "black",
      backgroundColor = TRUE,
      borderColor = TRUE
    ),
    striped = TRUE
  )

}
