% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_css.R
\name{tooltip_css}
\alias{tooltip_css}
\title{tooltipcss}
\usage{
tooltip_css(
  font_family = NULL,
  font_size = NULL,
  background_color = NULL,
  text_color = NULL
)
}
\arguments{
\item{font_family}{Font family for text in tooltips (default: "Public Sans")}

\item{font_size}{Font size for text in tooltips (default: "11px")}

\item{background_color}{Hover background colour (default: #22272b)}

\item{text_color}{Hover text colour (default: #F2F2F2)}
}
\value{
A character vector containing CSS styles for tooltips
}
\description{
CSS code required to add waratah-styled tooltips to interactive graphs created using ggiraph
}
