% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waratah-package.R
\docType{package}
\name{waratah-package}
\alias{waratah}
\alias{waratah-package}
\title{Tools for the New South Wales Design System in R}
\description{
This package allows NSW Government R users to create data visualisations in keeping with the NSW Design System. This ensures a consistent approach to data visualisations with appropriate colours and fonts. This package was inspired by the data viz work of Cara Thompson \url{https://github.com/cararthompson}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/digitalnsw/nsw-r-visualisations}
  \item Report bugs at \url{https://github.com/digitalnsw/nsw-r-visualisations/issues}
}

}
\author{
\strong{Maintainer}: Zoe Baldwin \email{zoe.baldwin@health.nsw.gov.au} (\href{https://orcid.org/0000-0002-6010-9361}{ORCID})

Other contributors:
\itemize{
  \item Carl Suster \email{carl.suster@sydney.edu.au} (\href{https://orcid.org/0000-0001-7021-9380}{ORCID}) [contributor]
}

}
